+++ example/example_basic/android/app/src/main/kotlin/com/example/example_basic/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.example_basic
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}