+++ example/example_basic/lib/models/user.dart
@@ -1,5 +1,0 @@
-class User {
-  final String firstName, lastName;
-
-  User(this.firstName, this.lastName);
-}